layer_1/requirements.txt
pandas==1.3.0
requests==2.28.2
beautifulsoup4==4.12.0
bs4==0.0.1
deepl==1.14.0
python-dotenv
lxml